solution to codebase through refactoring everything out of app.js and into views of collections
