+++ src/app.js
@@ -12,44 +12,6 @@ import TaskListView from './views/task_list_view';
 
 const taskList = new TaskList();
 
-// helper method for updating the DOM with the status from a hash
-const updateStatusMessageFrom = (messageHash) => {
-  $('#status-messages').empty();
-  _.each(messageHash, (messageType) => {
-    messageType.forEach((message) => {
-      $('#status-messages').append($(`<li>${message}</li>`));
-    })
-  });
-  $('#status-messages').show();
-}
-
-// helper method for updating the DOM with the status from a string
-const updateStatusMessageWith = (message) => {
-  $('#status-messages').empty();
-  $('#status-messages').append($(`<li>${message}</li>`));
-  $('#status-messages').show();
-}
-
-const addNewTask = function(event) {
-  event.preventDefault();
-
-  const taskData ={};
-  ['task_name', 'assignee'].forEach( (field) => {
-    const val = $(`input[name=${field}]`).val();
-    if (val != '') {
-      taskData[field] = val;
-    }
-  });
-  const newTask = new Task(taskData);
-
-  if (newTask.isValid()) {
-    taskList.add(newTask);
-    updateStatusMessageWith(`New task added: ${newTask.get('task_name')}`);
-  } else {
-    updateStatusMessageFrom(newTask.validationError);
-  }
-}
-
 $(document).ready( () => {
   const taskListView = new TaskListView({
     model: taskList,
@@ -58,8 +20,6 @@ $(document).ready( () => {
   });
 
   taskListView.render();
-
-  $('#add-task-form').submit(addNewTask);
 
   taskList.add(new Task({task_name: "Put rendering logic in Backbone Views", assignee: "Me"}));
   taskList.add(new Task({task_name: "Put handling events in Backbone Views", assignee: "Me"}));
