+++ src/css/style.css
@@ -19,7 +19,8 @@ body, h1, h2, h3, p {
 }
 
 .task {
-  padding: 30px 50px;
+  margin-bottom: 20px;
+  padding: 30px;
   border-radius: 5px;
 }
